Make harness logging minimal by default

+++ missions/cmgr_missions/cmgr_motion/zlaunch.sh
@@ -13,6 +13,7 @@ TIME_WARP=10
 MAX_TIME=60
 VERBOSE=no
 JUST_MAKE=no
+LOG_MODE=minimal
 DISPLAY_ARGS=(--nogui)
 FLOW_ARGS=()
 
@@ -24,6 +25,7 @@ Options:
   --help, -h           Show this help message
   --verbose, -v        Verbose launch output
   --just_make, -j      Only create targ files
+  --log=<mode>         Logging mode: minimal (default) or full
   --nogui, -ng         Headless launch, no gui (default)
   --gui                Launch with pMarineViewer
   --max_time=<secs>    Maximum time passed to xlaunch
@@ -52,6 +54,7 @@ for arg in "$@"; do
         --help|-h) usage; exit 0 ;;
         --verbose|-v) VERBOSE=yes; FLOW_ARGS+=(--verbose) ;;
         --just_make|-j) JUST_MAKE=yes ;;
+        --log=*) LOG_MODE="${arg#--log=}" ;;
         --nogui|-ng) DISPLAY_ARGS=(--nogui) ;;
         --gui) DISPLAY_ARGS=() ;;
         --max_time=*) MAX_TIME="${arg#--max_time=}" ;;
@@ -67,6 +70,10 @@ done
 
 is_uint "$TIME_WARP" || die "time warp must be an integer"
 is_uint "$MAX_TIME" || die "--max_time must be an integer"
+case "$LOG_MODE" in
+    minimal|full) ;;
+    *) die "--log must be minimal or full" ;;
+esac
 
 REPO_DIR=$(git -C "$PWD" rev-parse --show-toplevel 2>/dev/null) || \
     die "unable to locate repository root from $PWD"
@@ -101,8 +108,14 @@ on_signal() {
 trap on_exit EXIT
 trap on_signal INT TERM
 
+if [ "${LOG_MODE_PREPARED:-no}" != yes ]; then
+    ./prepare_logging_mode.sh "$LOG_MODE" || die "unable to prepare --log=$LOG_MODE"
+fi
+export LOG_MODE_PREPARED=yes
+export LOG_MODE_PREPARED_VALUE="$LOG_MODE"
+
 : > results.txt
-[ "$VERBOSE" = yes ] && echo "$ME: launching with max_time=$MAX_TIME warp=$TIME_WARP"
+[ "$VERBOSE" = yes ] && echo "$ME: launching with max_time=$MAX_TIME warp=$TIME_WARP log=$LOG_MODE"
 
 if [ "$JUST_MAKE" = yes ]; then
     rm -f targ_shoreside.moos targ_abe.moos targ_abe.bhv
